[Refactor] Shifted some exception class to temp folder.
