+++ src/main/java/controllers/temp/ConsoleInputController.java
@@ -1,5 +1,7 @@
-package controllers;
+package controllers.temp;
 
+import controllers.IViewController;
+import controllers.ProjectInputController;
 import exceptions.DukeException;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
