+++ src/main/java/controllers/temp/RecurringFactory.java
@@ -1,4 +1,4 @@
-package controllers;
+package controllers.temp;
 
 import exceptions.DukeException;
 import models.tasks.Recurring;
