+++ src/main/java/exceptions/temp/NoCommandDetailsException.java
@@ -1,4 +1,6 @@
-package exceptions;
+package exceptions.temp;
+
+import exceptions.DukeException;
 
 public class NoCommandDetailsException extends DukeException {
     public NoCommandDetailsException() {
