+++ src/main/java/exceptions/temp/NoTaskDetailsException.java
@@ -1,4 +1,6 @@
-package exceptions;
+package exceptions.temp;
+
+import exceptions.DukeException;
 
 public class NoTaskDetailsException extends DukeException {
