+++ src/test/java/DukeTest.java
@@ -1,4 +1,4 @@
-import controllers.TaskFactory;
+import controllers.temp.TaskFactory;
 import exceptions.DukeException;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
