+++ src/test/java/TodoTest.java
@@ -1,4 +1,4 @@
-import controllers.TaskFactory;
+import controllers.temp.TaskFactory;
 import exceptions.DukeException;
 import models.tasks.ITask;
 import models.tasks.ToDos;
